Reduce the number of modules under `bevy_render_2d::material`
