--- crates/bevy_render_2d/src/material/alpha_mode.rs
@@ -1,27 +1,0 @@
-use bevy_reflect::{prelude::ReflectDefault, Reflect};
-
-/// Sets how a 2d material's base color alpha channel is used for transparency.
-/// Currently, this only works with [`Mesh2d`](bevy_render::mesh::Mesh2d). Sprites are always transparent.
-///
-/// This is very similar to [`AlphaMode`](bevy_render::alpha::AlphaMode) but this only applies to 2d meshes.
-/// We use a separate type because 2d doesn't support all the transparency modes that 3d does.
-#[derive(Debug, Default, Reflect, Copy, Clone, PartialEq)]
-#[reflect(Default, Debug, Clone)]
-pub enum AlphaMode2d {
-    /// Base color alpha values are overridden to be fully opaque (1.0).
-    #[default]
-    Opaque,
-    /// Reduce transparency to fully opaque or fully transparent
-    /// based on a threshold.
-    ///
-    /// Compares the base color alpha value to the specified threshold.
-    /// If the value is below the threshold,
-    /// considers the color to be fully transparent (alpha is set to 0.0).
-    /// If it is equal to or above the threshold,
-    /// considers the color to be fully opaque (alpha is set to 1.0).
-    Mask(f32),
-    /// The base color alpha value defines the opacity of the color.
-    /// Standard alpha-blending is used to blend the fragment's color
-    /// with the color behind it.
-    Blend,
-}
